% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/einfach_shiny.R
\name{einfach}
\alias{einfach}
\title{GUI frontend for collecting tweets}
\usage{
einfach(data_path = NULL, verbose = FALSE)
}
\arguments{
\item{data_path}{path for storing your data; default to a temporatory directory}

\item{verbose}{if \code{FALSE}, no output}
}
\value{
Nothing
}
\description{
This function launches a GUI frontend for collecting tweets using the Twitter Academic Research Product Track v2 API endpoint. Please use \code{set_bearer} to setup your bearer token first.
}
