% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/einfach_shiny.R
\name{get_bearer}
\alias{get_bearer}
\title{Reexport get_bearer from academictwitteR}
\usage{
get_bearer()
}
\value{
your bearer token, if it has been setup.
}
\description{
This is a reexport of the function get_bearer from academictwitteR. If you want to know how to setup your access token, see \code{?academictwitteR::get_bearer}
}
