% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/einfach_shiny.R
\name{einfach}
\alias{einfach}
\title{GUI frontend for collecting tweets}
\usage{
einfach(data_path = NULL, bearer_token_path = "~/.academictwitteR_token")
}
\arguments{
\item{data_path}{path for storing your data; default to a temporatory directory}

\item{bearer_token_path}{path to your bearer token. Please setup using \code{set_bearer}}
}
\value{
Nothing
}
\description{
This function launches a GUI frontend for collecting tweets using the Twitter Academic Research Product Track v2 API endpoint. Please use \code{set_bearer} to setup your bearer token first.
}
